DFR-3123: Refactored radio enums to support extensibility more effectively
