+++ playwright-e2e/functional/contested/create_case/create_case.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect } from '../../../fixtures/fixtures';
 import { createCaseInCcd } from '../../../../test/helpers/utils';
 import config from '../../../config/config';
-import { RadioEnum } from '../../../pages/helpers/enums/RadioEnum';
+import { YesNoRadioEnum } from '../../../pages/helpers/enums/RadioEnums';
 import { createCaseTabData } from '../../../data/tab_content/contested/solicitor_create_case_tabs';
 
 test(
@@ -59,7 +59,7 @@ test(
 
     //applicant details
     const keepPrivate: boolean = true;
-    const applicantInRefuge: RadioEnum = RadioEnum.YES;
+    const applicantInRefuge: YesNoRadioEnum = YesNoRadioEnum.YES;
     await applicantDetailsPage.enterApplicantDetailsContested('Frodo', 'Baggins', keepPrivate, applicantInRefuge);
     await applicantDetailsPage.navigateContinue();
