+++ playwright-e2e/functional/contested/create_case/create_papercase_caseworker.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect } from '../../../fixtures/fixtures';
 import { createCaseInCcd } from '../../../../test/helpers/utils';
 import config from '../../../config/config';
-import { RadioEnum } from '../../../pages/helpers/enums/RadioEnum';
+import { YesNoRadioEnum, ApplicationtypeEnum } from '../../../pages/helpers/enums/RadioEnums';
 import {createCaseTabData} from "../../../data/tab_content/contested/caseworker_create_case_tabs";
 
 // Create a test case for the Contested Paper Case
@@ -56,8 +56,8 @@ test(
     await solicitorDetailsPage.enterReferenceNumber('Y707HZM');
     await solicitorDetailsPage.enterUKaddress();
     // Check both application types are present.
-    await solicitorDetailsPage.selectApplicationType(RadioEnum.CHILDRENS_ACT);
-    await solicitorDetailsPage.selectApplicationType(RadioEnum.MARRIAGE_CIVIL);
+    await solicitorDetailsPage.selectApplicationType(ApplicationtypeEnum.CHILDRENS_ACT);
+    await solicitorDetailsPage.selectApplicationType(ApplicationtypeEnum.MARRIAGE_CIVIL);
     await solicitorDetailsPage.navigateContinue();
 
     // Enter Divorce / Dissolution Details
@@ -66,7 +66,7 @@ test(
 
     //applicant details
     const keepPrivate: boolean = true;
-    const applicantInRefuge: RadioEnum = RadioEnum.YES;
+    const applicantInRefuge: YesNoRadioEnum = YesNoRadioEnum.YES;
     await applicantDetailsPage.enterApplicantDetailsContested('Frodo', 'Baggins', keepPrivate, applicantInRefuge);
     await applicantDetailsPage.navigateContinue();
