--- playwright-e2e/pages/helpers/enums/RadioEnum.ts
@@ -1,7 +1,0 @@
-export enum RadioEnum {
-    YES = 'Yes',
-    NO = 'No',
-    MARRIAGE_CIVIL = 'In connection to matrimonial and civil partnership proceedings (divorce/dissolution etc)',
-    CHILDRENS_ACT = 'Under paragraph 1 or 2 of schedule 1 children act 1989',
-    BLANK = '',
-  }
