+++ playwright-e2e/pages/helpers/enums/RadioEnums.ts
@@ -1,0 +1,10 @@
+export enum YesNoRadioEnum {
+    YES = 'Yes',
+    NO = 'No',
+    BLANK = '',
+  }
+
+export enum ApplicationtypeEnum {
+  MARRIAGE_CIVIL = 'In connection to matrimonial and civil partnership proceedings (divorce/dissolution etc)',
+  CHILDRENS_ACT = 'Under paragraph 1 or 2 of schedule 1 children act 1989',
+}
